✏️ [fix] UserService 인터페이스 구조로 변경
