+++ src/main/java/com/likelion/tostar/domain/user/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.likelion.tostar.domain.user.controller;
 
-import com.likelion.tostar.domain.user.dto.JoinRequestDTO;
+import com.likelion.tostar.domain.user.dto.UserInfoDTO;
 import com.likelion.tostar.domain.user.dto.LoginRequestDTO;
-import com.likelion.tostar.domain.user.service.UserService;
+import com.likelion.tostar.domain.user.service.UserServiceImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -14,22 +14,22 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/user")
 @RequiredArgsConstructor
-public class MemberController {
-    private final UserService userService;
+public class UserController {
+    private final UserServiceImpl userServiceImpl;
 
     /**
      * 회원 로그인
      */
     @PostMapping("/login")
     public ResponseEntity<?> login(@Valid @RequestBody LoginRequestDTO loginRequestDTO) {
-        return userService.login(loginRequestDTO);
+        return userServiceImpl.login(loginRequestDTO);
     }
 
     /**
      * 회원 가입
      */
     @PostMapping("/join")
-    public ResponseEntity<?> join(@Valid @RequestBody JoinRequestDTO joinRequestDTO) {
-        return userService.join(joinRequestDTO);
+    public ResponseEntity<?> join(@Valid @RequestBody UserInfoDTO userInfoDTO) {
+        return userServiceImpl.join(userInfoDTO);
     }
 }
